Fix: 타입변경으로 인한 에러 수정
- 구현 중이던 코드베이스에 Integer -> Long 타입변경 적용하면서 생긴 충돌 수정
- 브랜치명이 잘못된 것을 확인하고 수정 (/deposit -> /payments)

+++ semicolon/src/main/java/dukku/semicolon/boundedContext/payment/app/ConfirmPaymentUseCase.java
@@ -52,8 +52,8 @@ public class ConfirmPaymentUseCase {
 
         // 5. 결제 승인 상태 변경 전 값 저장
         PaymentStatus originStatus = payment.getPaymentStatus();
-        Integer originAmountPg = payment.getAmountPg();
-        Integer originDeposit = payment.getPaymentDeposit();
+        Long originAmountPg = payment.getAmountPg();
+        Long originDeposit = payment.getPaymentDeposit();
 
         // 6. 결제 승인 처리
         // TODO: 실제 토스 API 호출
@@ -75,7 +75,7 @@ public class ConfirmPaymentUseCase {
         }
     }
 
-    private void validateAmount(Payment payment, Integer tossAmount) {
+    private void validateAmount(Payment payment, Long tossAmount) {
         if (!payment.getAmountPg().equals(tossAmount)) {
             throw new TossAmountMismatchException(payment.getAmountPg(), tossAmount);
         }
@@ -88,7 +88,7 @@ public class ConfirmPaymentUseCase {
     }
 
     private void createHistory(Payment payment, PaymentStatus originStatus,
-            Integer originAmountPg, Integer originDeposit) {
+            Long originAmountPg, Long originDeposit) {
         PaymentHistory history = PaymentHistory.create(
                 payment,
                 PaymentHistoryType.PAYMENT_SUCCESS,
